+++ src/object.h
@@ -37,6 +37,6 @@ PyObject* Object_get_oid(Object *self);
 PyObject* Object_get_hex(Object *self);
 PyObject* Object_get_type(Object *self);
 PyObject* Object_read_raw(Object *self);
-PyObject* wrap_object(git_object *c_object, Repository *repo);
+PyObject* wrap_object(git_object *c_object, Repository *repo, const git_tree_entry *entry);
 
 #endif